requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
httpx
pytest
asyncpg
pytest-asyncio
aiosqlite